src/RMBG-2.0-demo/requirements.txt
torch
torchvision
pillow
kornia
transformers
streamlit
huggingface
timm
modelscope
psutil